codalab/requirements/common.txt
alabaster==0.7.10
alembic==0.6.7
amqp==2.2.2
anyjson==0.3.3
argh==0.26.2
azure==0.11.1
Babel==2.5.1
billiard==3.5.0.3
boto==2.45.0
cachetools==2.0.1
celery==4.1.0
certifi==2017.11.5
chardet==3.0.4
closure-linter==2.3.19
colorama==0.3.9
coverage==4.4.2
dill==0.2.7.1
Django==1.8.5
#django-allauth==0.11.1
django-allauth
django-analytical==2.4.0
django-appconf==0.6
django-bootstrap-form==3.2
django-braces==1.2.2
django-compressor==2.2
django-config-gen==1.0.8
django-configurations==2.0
#django-debug-panel==0.8.1
django-debug-panel
#django-debug-toolbar==1.3.0
django-debug-toolbar
django-extensions==1.5.7
django-extra-views==0.6.3
django-filter==1.1.0
django-guardian==1.4.9
django-haystack==2.4.0
django-js-reverse==0.8.1
django-mptt==0.8.6
django-nose==1.4.1
django-oauth-toolkit==0.6.1
django-pin-passcode==0.1.9
# django-s3direct==0.4.8
django-s3direct
django-simple-captcha==0.4.5
django-storages==1.6.5
django-tinymce==2.0.5
django-userswitch==0.2.2
djangorestframework==2.3.6
docopt==0.6.2
docutils==0.14
enum34==1.1.6
future==0.16.0
futures==3.2.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
google-api-core==0.1.4
google-auth==1.3.0
google-cloud==0.32.0
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.0
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.0
google-cloud-core==0.28.0
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.0
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.0
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.0
google-cloud-runtimeconfig==0.28.0
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.0
google-cloud-videointelligence==1.0.0
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
grpc-google-iam-v1==0.11.4
grpcio==1.8.4
gunicorn==19.7.0
httplib2==0.10.3
idna==2.6
imagesize==0.7.1
Jinja2==2.10
kombu==4.1.0
Mako==1.0.7
MarkupSafe==1.0
mock==1.0.1
newrelic==2.54.0.41
nose==1.3.7
oauth2client==3.0.0
oauthlib==0.6.1
pathtools==0.1.2
Pillow==5.0.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.5.1
psutil==2.1.1
psycopg2==2.7.1
py==1.5.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
Pygments==2.2.0
pyrabbit==1.1.0
pytest==3.0.6
pytest-django==2.9.1
pytest-testmon==0.9.6
pytest-watch==4.1.0
python-dateutil==2.6.1
python-gflags==3.1.2
python-memcached==1.57
python-openid==2.2.5
python-ptrace==0.6.5
pytz==2013b0
PyYAML==3.10
rcssmin==1.0.6
requests==2.18.4
requests-oauthlib==0.3.2
rjsmin==1.0.12
rsa==3.4.2
six==1.10.0
snowballstemmer==1.2.1
Sphinx==1.4.1
SQLAlchemy==0.9.7
#sqlparse==0.1.19
sqlparse
urllib3==1.22
vine==1.1.4
virtualenv==15.1.0
watchdog==0.8.1